requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP and Text Processing
scikit-learn==1.3.2

# Notifications
sendgrid==6.11.0
slack-sdk==3.26.1

# Security
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2

# OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Testing
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
